[feat] Move Everything into Component Class
